organize sepana lib dir
