+++ src/pages/api/sepana/deleteAll.page.ts
@@ -1,19 +1,19 @@
-import { deleteAllRecords } from 'lib/sepana'
-import { sepanaAlert } from 'lib/sepana/log'
+import { deleteAllRecords } from 'lib/sepana/api'
+import { sepanaLog } from 'lib/sepana/log'
 import { NextApiHandler } from 'next'
 
 const handler: NextApiHandler = async (_, res) => {
   try {
     await deleteAllRecords()
 
-    await sepanaAlert({ type: 'alert', alert: 'DELETED_ALL_RECORDS' })
+    await sepanaLog({ type: 'alert', alert: 'DELETED_ALL_RECORDS' })
 
     res.status(200).json({
       network: process.env.NEXT_PUBLIC_INFURA_NETWORK,
       message: 'Deleted all Sepana records',
     })
   } catch (error) {
-    await sepanaAlert({
+    await sepanaLog({
       type: 'alert',
       alert: 'DB_UPDATE_ERROR',
       body: JSON.stringify(error),
